create abstract class for test rule widget
