+++ browser-extension/common/src/custom-elements/index.ts
@@ -1,4 +1,4 @@
 import "./toast";
 import "./session-recording-widgets/manual-mode-widget";
 import "./session-recording-widgets/auto-mode-widget";
-import "./test-rule-widget";
+import "./test-rule-widget/explicit-test-rule-widget";
